requirements.txt
streamlit
anthropic
pillow
bcrypt
pyyaml
instabot
pyperclip